refactor(tasks): move the Spider family into sieval/tasks/spider/
`sieval/tasks/CLAUDE.md` puts the threshold at five files per benchmark, and
Spider 1.0 is exactly there: one task module plus four private helpers
(`_spider_exec`, `_spider_schema`, `_spider_sqlite`, `_spider_test_suite`).
It is also already the flat outlier -- every other benchmark in `tasks/` has at
most ONE private helper, so four is what the rule means by logic a flat layout
has nowhere good to put.

Matches the two existing subpackages (`arc/`, `platinum_bench/`) exactly: an
**empty** `__init__.py`, because both resolution mechanisms key on the module
rather than a package attribute, so there stays one import path per task and no
type surface to drift. `tests/unit/tasks/spider/` mirrors it, with the same
empty `__init__.py` `tests/unit/tasks/arc/` carries.

Everything inside the package keeps its relative imports, so only absolute
references moved -- 16 of them, across the five tests, the import-discipline
manifest and two docstring cross-references. Two were the kind a move breaks
quietly: the manifest is keyed by module path under `sieval.tasks.`, so the key
becomes `spider.spider_0shot_gen` or the probe examines a module that no longer
exists; and two `monkeypatch.setattr` targets are STRINGS naming the old path,
which stop intercepting rather than erroring if the old path still resolves.

Verified against the failure modes a move of this shape actually has:

* **Name resolution.** `sieval task show spider_0shot_gen` still resolves and
  reports `Ready: yes`. This is the one CLAUDE.md warns about -- with only the
  lazy-export half working, a nested task imports fine but is unresolvable by
  name, and the only symptom is a bare `KeyError`.
* **Fix carry-over, by reverse-mutation at the NEW location.** Both review fixes
  are still load-bearing there: un-masking comments in `_statement_from` turns
  `test_extract_ignores_a_keyword_inside_a_leading_comment` red, and restoring
  `if not extra.get("parsed")` turns both `n_parser_rejected` tests red. A move
  that silently reverted either would have left these mutations passing.
* 155 Spider-and-neighbour tests, 6,667 unit+integration, 25/25 preflight,
  `ruff`/`ty` clean. `[tool.mutmut] also_copy` already lists `sieval/tasks`, so
  the new subpackage needs no config change, and `meta/index.json` is unchanged
  -- it records task metadata, not module paths.

Scope: only Spider 1.0's files exist on this branch. #125 adds
`spider2_lite_0shot_gen` and `_spider2_backends` and modifies two of the helpers
above, so it should land those inside this package when it rebases -- which is
what makes the directory earn its keep, since the helpers then genuinely serve
two variants.
