+++ sieval/tasks/__init__.pyi
@@ -190,7 +190,7 @@ from .simpleqa_verified_0shot_gen import (
 from .smt_2025_0shot_gen import (
     SMT2025ZeroShotGenTask,
 )
-from .spider_0shot_gen import (
+from .spider.spider_0shot_gen import (
     SpiderZeroShotGenTask,
 )
 from .sysbench_0shot_gen import (
